email_sender/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
openai>=1.0.0
openai-agents>=0.1.0
fastapi>=0.111.0
uvicorn>=0.30.0

# Email service
sendgrid>=6.10.0

# HTTP requests
requests>=2.31.0

# Async support
asyncio

# Type hints
typing-extensions>=4.0.0
